Extract generator API
