+++ core/build.gradle.kts
@@ -1,0 +1,9 @@
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(project(":generators-api"))
+            }
+        }
+    }
+}
