+++ core/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/Enums.kt
@@ -1,7 +1,5 @@
 package com.github.jcornaz.kwik.generator
 
-import com.github.jcornaz.kwik.Generator
-
 /**
  * Returns a [Generator] emitting random enum instance picked in the enum [T]
  */
