+++ core/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/Strings.kt
@@ -1,6 +1,5 @@
 package com.github.jcornaz.kwik.generator
 
-import com.github.jcornaz.kwik.Generator
 import com.github.jcornaz.kwik.filterNot
 import com.github.jcornaz.kwik.withSamples
