+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/CreateGeneratorTest.kt
@@ -1,6 +1,5 @@
 package com.github.jcornaz.kwik.generator
 
-import com.github.jcornaz.kwik.Generator
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
