+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/EnumGeneratorTest.kt
@@ -1,7 +1,6 @@
 package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.AbstractGeneratorTest
-import com.github.jcornaz.kwik.Generator
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
