+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/operator/FilterTest.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.operator
 
 import com.github.jcornaz.kwik.*
+import com.github.jcornaz.kwik.generator.Generator
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
