+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/operator/MapTest.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.operator
 
 import com.github.jcornaz.kwik.AbstractGeneratorTest
-import com.github.jcornaz.kwik.Generator
+import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.map
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
