+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/runner/ForAll4Test.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.runner
 
 import com.github.jcornaz.kwik.FalsifiedPropertyError
-import com.github.jcornaz.kwik.Generator
+import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.forAll
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
