+++ settings.gradle.kts
@@ -1,3 +1,5 @@
 rootProject.name = "kwik"
 
 include("core")
+
+include("generators-api")
